Add files via upload
Updated to mainUI so that it now mostly functions as desired, updating dropdown options when changes are made to the game mode and profession selections. Also added a button to save the all of the selections to the Test.db file. Changes were made to the Test.db to remove a few tables that I felt were unnecessary. Still think I can make the mainUI better breaking a lot of it into smaller class frames and going to continue looking into how to do that.